+++ src/Phan/AST/UnionTypeVisitor.php
@@ -1763,10 +1763,11 @@ class UnionTypeVisitor extends AnalysisVisitor
      *
      * @throws IssueException
      * if the unpack is on an invalid expression
+     * @suppress PhanUndeclaredProperty
      */
     public function visitUnpack(Node $node) : UnionType
     {
-        return $this->analyzeUnpack($node, false);
+        return $this->analyzeUnpack($node, isset($node->is_in_array));
     }
 
     /**