+++ src/Phan/Language/Type/StringType.php
@@ -46,7 +46,7 @@ class StringType extends ScalarType
         if ($other instanceof ScalarType) {
             return $other instanceof StringType || (!$context->isStrictTypes() && parent::canCastToDeclaredType($code_base, $context, $other));
         }
-        return $other instanceof CallableType;
+        return $other instanceof CallableType || $other instanceof MixedType;
     }
 
     /**