+++ tests/files/src/0795_array_iteration.php
@@ -1,0 +1,8 @@
+<?php
+namespace NS795\SubNS;
+/** @param array $points */
+function test($points) {
+    foreach ($points as $point) {
+        echo $point[1];
+    }
+}